Split off implementation details that shouldn't be visible to the user.
